refactor(exception): 에러 코드 수정
