+++ src/main/java/woowacourse/shoppingcart/cart/exception/badrequest/InvalidQuantityException.java
@@ -1,10 +1,10 @@
-package woowacourse.shoppingcart.customer.exception.badrequest;
+package woowacourse.shoppingcart.cart.exception.badrequest;
 
 import woowacourse.shoppingcart.exception.BadRequestException;
 
 public class InvalidQuantityException extends BadRequestException {
 
     public InvalidQuantityException() {
-        super("999", "수량이 유효하지 않습니다.");
+        super("1000", "수량이 유효하지 않습니다.");
     }
 }
