+++ src/main/java/woowacourse/shoppingcart/customer/dao/CustomerDao.java
@@ -1,7 +1,6 @@
 package woowacourse.shoppingcart.customer.dao;
 
 import java.sql.PreparedStatement;
-import java.util.Locale;
 import java.util.Objects;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
@@ -9,7 +8,6 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.stereotype.Repository;
 import woowacourse.shoppingcart.customer.domain.Customer;
-import woowacourse.shoppingcart.customer.exception.badrequest.InvalidCustomerException;
 import woowacourse.shoppingcart.customer.exception.notfound.NotFoundCustomerException;
 
 @Repository
@@ -40,15 +38,6 @@ public class CustomerDao {
         }, keyHolder);
 
         return Objects.requireNonNull(keyHolder.getKey()).longValue();
-    }
-
-    public Long findInByNickname(final String nickname) {
-        try {
-            final String query = "SELECT id FROM customer WHERE nickname = ?";
-            return jdbcTemplate.queryForObject(query, Long.class, nickname.toLowerCase(Locale.ROOT));
-        } catch (final EmptyResultDataAccessException e) {
-            throw new InvalidCustomerException();
-        }
     }
 
     public boolean existEmail(final String email) {
