+++ src/main/java/woowacourse/shoppingcart/customer/exception/badrequest/InvalidEmailException.java
@@ -5,6 +5,6 @@ import woowacourse.shoppingcart.exception.BadRequestException;
 public class InvalidEmailException extends BadRequestException {
 
     public InvalidEmailException() {
-        super("997", "이메일이 유효하지 않습니다.");
+        super("1000", "이메일이 유효하지 않습니다.");
     }
 }
