+++ src/main/java/woowacourse/shoppingcart/customer/exception/badrequest/InvalidNicknameException.java
@@ -5,6 +5,6 @@ import woowacourse.shoppingcart.exception.BadRequestException;
 public class InvalidNicknameException extends BadRequestException {
 
     public InvalidNicknameException() {
-        super("997", "닉네임이 유효하지 않습니다.");
+        super("1000", "닉네임이 유효하지 않습니다.");
     }
 }
