+++ src/main/java/woowacourse/shoppingcart/customer/exception/badrequest/InvalidPasswordException.java
@@ -5,6 +5,6 @@ import woowacourse.shoppingcart.exception.BadRequestException;
 public class InvalidPasswordException extends BadRequestException {
 
     public InvalidPasswordException() {
-        super("997", "비밀번호가 유효하지 않습니다.");
+        super("1000", "비밀번호가 유효하지 않습니다.");
     }
 }
