+++ src/main/java/woowacourse/shoppingcart/customer/exception/notfound/NotFoundCustomerException.java
@@ -5,6 +5,6 @@ import woowacourse.shoppingcart.exception.notfound.NotFoundException;
 public class NotFoundCustomerException extends NotFoundException {
 
     public NotFoundCustomerException() {
-        super("999", "회원이 존재하지 않습니다.");
+        super("2000", "회원이 존재하지 않습니다.");
     }
 }
